pyproject.toml
[project]
name            = "refuel-automation"
version         = "0.1.0"
description     = "Refuel automation"
readme          = "README.md"
requires-python = ">=3.12"

# Requirements
dependencies = [
    "vyper==0.4.3",
    "snekmate==0.1.2",
    "titanoboa==0.2.8",
    "web3>=7.14.0",
]

[tool.uv]
dev-dependencies = [
    "pytest-xdist>=3.8.0",
    "mamushi==0.1.0",
    "pre-commit==4.5.1",
    "ruff==0.14.13",
    "pytest==9.0.2",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 100
